add another website ip quire result handler

+++ src/com/tc/HostMode.java
@@ -5,5 +5,5 @@ package com.tc;
  */
 public class HostMode {
     String ip;
-    int timeMs;
+    double timeMs;
 }
